Computers are fast Qemu is not. Also volatile
Taped on enough volatile statements to make the the test pattern work under -Ofast
The emulation of risc-v is radically faster than the serial out. Modifying qemu source
is 100% on the table for the future. The next step however will have to be system calls
then loading programs of the disk. This will enable me to investigate the best API design.