Drive partition files
kmain now calls load_drive_partitions() that loads the partition table on
drive1.dsk and creates an imaginary directory stored in the drive1_partition_directory
variable. This directory is then populated with drive partition files that represent
the partitions on drive1. Reading and writing is not yet implemented for these files.