Multicore and statistics
This commit contains bug fixes and performance improvements to multicore.
It also has added statistics related to the time lost to locking against
the process array. Most syscalls don't actually need to get a write lock though
as they currently are doing. This is because locking as write guarantees exclusive
access. As I move more and more operations onto more locally scoped locks we
can move the ops onto "read" locking the proc array only. And with this we should
see the time lost to locking the proc array, *currently 20%!, go down.